fix(neighbors): enhance batch tile buffer estimation for empty inputs
Update the JAX implementation to handle empty batches gracefully in the _batch_tile_buffer_max_tiles_per_group function. Introduce validation for cell_batch dimensions in the estimate_batch_max_tiles_per_group function. Modify the Torch implementation to ensure determinant calculations use float64 for float32 cell inputs, improving accuracy in volume estimations. Add corresponding unit tests for these changes.

+++ nvalchemiops/torch/neighbors/batch_cluster_tile.py
@@ -147,7 +147,7 @@ def estimate_batch_max_tiles_per_group(
     int
         Shared ``max_tiles_per_group`` for the batched compact tile buffer.
     """
-    volumes = torch.linalg.det(cell_batch).abs().view(-1)
+    volumes = torch.linalg.det(cell_batch.to(torch.float64)).abs().view(-1)
     return wp_estimate_batch_max_tiles_per_group(
         batch_ptr,
         cutoff,
